[ALPS03597512] Sensor: Fix build error
[Detail]
Cast to correct size

MTK-Commit-Id: 821bf80e4682b402c8c5d93fb1ed2270d5faa451

CR-Id: ALPS03597512
Feature: MT6771 Platform Feature

+++ drivers/misc/mediatek/sensors-1.0/sensorHub/SCP_nanoHub/SCP_nanoHub.c
@@ -1866,7 +1866,7 @@ static int sensorHub_power_up_work(void *data)
 		obj->wp_queue.head = 0;
 		obj->wp_queue.tail = 0;
 		/* 2. init dram information */
-		obj->SCP_sensorFIFO = (struct sensorFIFO *)scp_get_reserve_mem_virt(SENS_MEM_ID);
+		obj->SCP_sensorFIFO = (struct sensorFIFO *)(long)scp_get_reserve_mem_virt(SENS_MEM_ID);
 		WARN_ON(obj->SCP_sensorFIFO == NULL);
 		obj->SCP_sensorFIFO->wp = 0;
 		obj->SCP_sensorFIFO->rp = 0;
